Refactor, better separation of concerns. Centralise map pin processing to allow future code improvements
