+++ Frodo/Program.cs
@@ -30,7 +30,7 @@ var mapper = new MappingService();
 
 var ai = new MapPinService(selenium, pinHelper);
 var clientExportFileGenerator = new ClientExportFileGenerator(dbLoader, mapper, pinHelper);
-var service = new DataSyncService(ai, selenium, pinHelper, dbLoader, mapper, clientExportFileGenerator);
+var service = new DataSyncService(ai, pinHelper, dbLoader, mapper, clientExportFileGenerator);
 service.ProcessFile();
 
 selenium.Stop();
