+++ Samwise/MainWindow.xaml
@@ -48,7 +48,7 @@
         <TextBox x:Name="txtUrl" Grid.Row="0" Grid.Column="1" Margin="2,2,2,2" TextWrapping="Wrap" Text="Any"/>
 
         <TextBlock Grid.Row="1" Grid.Column="0" Margin="21,10,0,0" TextWrapping="Wrap" Text="Facebook" />
-        <Button x:Name="butFacebook" Content="Facebook" Grid.Row="1" Grid.Column="1" Style="{StaticResource Link}" Margin="0,5,0,6" VerticalAlignment="Stretch" Click="butFacebook_Click" />
+        <Button x:Name="butFacebook" Content="Facebook" Grid.Row="1" Grid.Column="1" Style="{StaticResource Link}" Margin="0,5,0,6" VerticalAlignment="Stretch" Click="ButFacebook_Click" />
 
         <TextBlock Grid.Row="2" Grid.Column="0" Margin="21,10,0,0" TextWrapping="Wrap" Text="Search Token" />
         <TextBox x:Name="txtSearchToken" Grid.Row="2" Grid.Column="1" Margin="2,2,2,2" TextWrapping="Wrap" Text=""/>
@@ -76,9 +76,9 @@
             <Grid.RowDefinitions>
                 <RowDefinition Height="27"/>
             </Grid.RowDefinitions>
-            <Button x:Name="butCaptureInfo" Content="Capture Info" Grid.Column="0" Grid.Row="0" HorizontalAlignment="Left" VerticalAlignment="Top" Click="butCaptureInfo_Click"   />
-            <Button x:Name="butLeft" Content="&lt;" Grid.Column="5" Grid.Row="0" VerticalAlignment="Top" Margin="0" HorizontalAlignment="Center" Width="22" Click="butLeft_Click"  />
-            <Button x:Name="butRight" Content="&gt;" Grid.Column="6" Grid.Row="0" HorizontalAlignment="Center" Margin="0" VerticalAlignment="Top" Width="22" Click="butRight_Click"  />
+            <Button x:Name="butCaptureInfo" Content="Capture Info" Grid.Column="0" Grid.Row="0" HorizontalAlignment="Left" VerticalAlignment="Top" Click="ButCaptureInfo_Click"   />
+            <Button x:Name="butLeft" Content="&lt;" Grid.Column="5" Grid.Row="0" VerticalAlignment="Top" Margin="0" HorizontalAlignment="Center" Width="22" Click="ButLeft_Click"  />
+            <Button x:Name="butRight" Content="&gt;" Grid.Column="6" Grid.Row="0" HorizontalAlignment="Center" Margin="0" VerticalAlignment="Top" Width="22" Click="ButRight_Click"  />
             <TextBlock x:Name="txtCounter" Grid.Row="0" Grid.Column="7" Margin="2,2,2,2" TextWrapping="Wrap" Text="1/ 100"/>
         </Grid>
