+++ _Library/Gluten.Core.DataProcessing/Gluten.Core.DataProcessing.csproj
@@ -20,8 +20,4 @@
     <ProjectReference Include="..\Gluten.Data\Gluten.Data.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Helper\" />
-  </ItemGroup>
-
 </Project>
