+++ gluten-angular-map/Update.ps1
@@ -1,3 +1,3 @@
 ng build
-swa deploy ./dist/gluten-angular-map/browser --env production --app-name GlutenAngular
+swa deploy ./dist/gluten-angular-map/browser --env production --app-name glutendev 
 pause
